move RefreshControl to components/core
