+++ src/screens/Tank/SettleBalance/SettleBalance.js
@@ -35,7 +35,7 @@ import Button from 'components/legacy/Button';
 import ListItemWithImage from 'components/legacy/ListItem/ListItemWithImage';
 import TankAssetBalance from 'components/TankAssetBalance';
 import Checkbox from 'components/legacy/Checkbox';
-import RefreshControl from 'components/RefreshControl';
+import RefreshControl from 'components/core/RefreshControl';
 import Spinner from 'components/Spinner';
 import Toast from 'components/Toast';
